+++ internal/controllers/gateway/gateway_controller_test.go
@@ -58,6 +58,110 @@ func Test_readyConditionExistsForObservedGeneration(t *testing.T) {
 	assert.False(t, isGatewayReady(neverBeenReadyGateway))
 }
 
+func Test_setGatewayCondtion(t *testing.T) {
+	testCases := []struct {
+		name            string
+		gw              *gatewayv1alpha2.Gateway
+		condition       metav1.Condition
+		conditionLength int
+	}{
+		{
+			name: "no_such_condition_should_append_one",
+			gw:   &gatewayv1alpha2.Gateway{},
+			condition: metav1.Condition{
+				Type:               "fake1",
+				Status:             metav1.ConditionTrue,
+				Reason:             "fake1",
+				ObservedGeneration: 1,
+				LastTransitionTime: metav1.Now(),
+			},
+			conditionLength: 1,
+		},
+		{
+			name: "have_condition_with_type_should_replace",
+			gw: &gatewayv1alpha2.Gateway{
+				Status: gatewayv1alpha2.GatewayStatus{
+					Conditions: []metav1.Condition{
+						{
+							Type:               "fake1",
+							Status:             metav1.ConditionFalse,
+							Reason:             "fake1",
+							ObservedGeneration: 1,
+							LastTransitionTime: metav1.Now(),
+						},
+					},
+				},
+			},
+			condition: metav1.Condition{
+				Type:               "fake1",
+				Status:             metav1.ConditionTrue,
+				Reason:             "fake1",
+				ObservedGeneration: 2,
+				LastTransitionTime: metav1.Now(),
+			},
+			conditionLength: 1,
+		},
+		{
+			name: "multiple_conditions_with_type_should_preserve_one",
+			gw: &gatewayv1alpha2.Gateway{
+				Status: gatewayv1alpha2.GatewayStatus{
+					Conditions: []metav1.Condition{
+						{
+							Type:               "fake1",
+							Status:             metav1.ConditionFalse,
+							Reason:             "fake1",
+							ObservedGeneration: 1,
+							LastTransitionTime: metav1.Now(),
+						},
+						{
+							Type:               "fake1",
+							Status:             metav1.ConditionTrue,
+							Reason:             "fake2",
+							ObservedGeneration: 2,
+							LastTransitionTime: metav1.Now(),
+						},
+						{
+							Type:               "fake2",
+							Status:             metav1.ConditionTrue,
+							Reason:             "fake2",
+							ObservedGeneration: 2,
+							LastTransitionTime: metav1.Now(),
+						},
+					},
+				},
+			},
+			condition: metav1.Condition{
+				Type:               "fake1",
+				Status:             metav1.ConditionTrue,
+				Reason:             "fake1",
+				ObservedGeneration: 3,
+				LastTransitionTime: metav1.Now(),
+			},
+			conditionLength: 2,
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			setGatewayCondition(tc.gw, tc.condition)
+			t.Logf("checking conditions of gateway after setting")
+			assert.Len(t, tc.gw.Status.Conditions, tc.conditionLength)
+
+			conditionNum := 0
+			var observedCondition metav1.Condition
+			for _, condition := range tc.gw.Status.Conditions {
+				if condition.Type == tc.condition.Type {
+					conditionNum++
+					observedCondition = condition
+				}
+			}
+			assert.Equal(t, 1, conditionNum)
+			assert.EqualValues(t, tc.condition, observedCondition)
+		})
+	}
+}
+
 func Test_isGatewayMarkedAsScheduled(t *testing.T) {
 	t.Log("verifying scheduled check for gateway object which has been scheduled")
 	scheduledGateway := &gatewayv1alpha2.Gateway{