+++ test/consts/gateway.go
@@ -1,5 +1,0 @@
-package consts
-
-const (
-	GatewayCRDsKustomizeURL = "github.com/kubernetes-sigs/gateway-api/config/crd/experimental?ref=v0.5.0"
-)